Commit Portfolio animé. Back-Office terminé.

+++ back/Projects.php
@@ -1,5 +1,15 @@
 <?php
+session_start();
 require '../Backoffice.php';
+$id = 0;
+$query = "SELECT idusers FROM users WHERE username='{$_SESSION["newsession"]}'";
+$req = $db->query($query);
+while($data = $req->fetch()){
+    $id = $data['idusers'];
+}
+if ($id = 0) {
+    die("Vous n'êtes pas autorisé à consulter cette page");
+}
 ?>
 <!DOCTYPE html>
 <html lang="fr">
@@ -23,7 +33,10 @@ require '../Backoffice.php';
         <div class="row min-vh-100">
             <div class="col-lg-2 coldeskLft"></div>
             <div class="col-lg-8 divForm d-flex flex-column justify-content-center">
-                <p class="titre">Projets</p>
+                <p class="titre">Gestion des Projets du Portfolio</p>
+                <div class="d-flex justify-content-end">
+                    <a href="addProject.php" class="bnt bntBottom">Ajouter un nouveau projet</a>
+                </div>
                 <table>
                     <tr>
                         <th>ID</th>
@@ -49,11 +62,11 @@ require '../Backoffice.php';
                         $datefr = date("d-m-Y H:m:s", strtotime($data['createdat']));
                         $ligne =  $ligne.  "<td>". $datefr. '</td>';
                         /* cinquième colonne : actions sur l'image */
-                        $ligne =  $ligne.  '<td>';
-                        $ligne =  $ligne. '<a href="editProject.php?id='. $data['idprojets']. '">Editer</a>';
-                        $ligne =  $ligne. '/';
-                        $ligne =  $ligne. '<a href="deleteProject.php?id='. $data['idprojets']. '">Supprimer</a>';
-                        $ligne =  $ligne. '</td>';
+                        $message = "Etes-vous sûr de vouloir supprimer l'image : ". $data['title'];
+                        $ligne =  $ligne.  '<td><div class="d-flex flex-column align-items-center justify-content-center gap-1">';
+                        $ligne =  $ligne. '<a href="editProject.php?id='. $data['idprojets']. '" class="bnt">Editer</a>';
+                        $ligne =  $ligne. '<a href="deleteProject.php?id='. $data['idprojets']. '" class="bnt" onclick="return confirm('. $message. ')">Supprimer</a>';
+                        $ligne =  $ligne. '</div></td>';
                          /* fin de ligne */
                          $ligne =  $ligne. "</tr>";
                           /* affichage de la ligne du tableau */
@@ -62,7 +75,14 @@ require '../Backoffice.php';
                 ?>
                 </table>
             </div>
-            <div class="col-lg-2 coldeskRght"></div>
+            <div class="col-lg-2 coldeskRght justify-content-center">
+                <a href="deconnect.php">
+                    <svg xmlns="http://www.w3.org/2000/svg" width="34" height="34" fill="currentColor" class="bi bi-door-closed" viewBox="0 0 16 16">
+                    <path d="M3 2a1 1 0 0 1 1-1h8a1 1 0 0 1 1 1v13h1.5a.5.5 0 0 1 0 1h-13a.5.5 0 0 1 0-1H3V2zm1 13h8V2H4v13z"/>
+                    <path d="M9 9a1 1 0 1 0 2 0 1 1 0 0 0-2 0z"/>
+                    </svg>
+                </a>
+            </div>
         </div>
 </div>
 </body>
